s2: camera: fix compilation errors

+++ camera/QCamera2/HAL/QCameraPostProc.cpp
@@ -2542,11 +2542,11 @@ void *QCameraPostProcessor::dataSaveRoutine(void *data)
                         if ((ssize_t)job_data->out_data.buf_filled_len != written_len) {
                             ALOGE("%s: Failed save complete data %ld bytes "
                                   "written instead of %d bytes!",
-                                  __func__, written_len,
+                                  __func__, (long)written_len,
                                   job_data->out_data.buf_filled_len);
                         } else {
                             CDBG_HIGH("%s: written number of bytes %ld\n",
-                                __func__, written_len);
+                                __func__, (long)written_len);
                         }
 
                         close(file_fd);
